refactor: update SortableList component to use generics
This commit updates the SortableList component to use generics, allowing for more flexibility in the data type of the list items. The SortableListProps interface now includes a generic type parameter T, which represents the type of the data array. The onChangeSortableList callback function also now accepts an array of type T. This change improves the reusability and type safety of the SortableList component.

Ref: ABC-123